* master: (263 commits)
  Fixed ISSET/ISEMPTY bit meaning to simplify run-time checks
  Use hrtime() for timing tests
  Fix AVX part
  Do one test build with AVX
  Export standard object handlers, to avoid indirect access
  Optimize zend_new_array() for special case, when size <= HT_MIN_SIZE
  Use SSE/AVX instructions
  Use SSE2 instruction to reset HashTable
  Avoid rehashing of "packed" array
  Eliminate some checks for zero character inside an inclue file name
  Avoid agressive inlining in rarely used paths
  Marked rarely used opcodes as "cold" (e.g. ADD_CONST_CONST ususaly optimized out).
  Remove build system files of bundled library libbcmath
  Revert "Avoid reference-counting"
  Removed useless zval_ptr_dtor()
  Fixed HashTable load factor
  Replace unused generic zend_hash_*add_or_update() functions by compact implmentations.
  _zend_hash_index_add_or_update_i() optimization
  Update year range in hrtime c and h files
  Remove unused binary_pipes option
  ...

+++ ext/phar/php_phar.h
@@ -34,7 +34,7 @@ extern zend_module_entry phar_module_entry;
 #define PHP_PHAR_API PHPAPI
 #endif
 
-PHP_PHAR_API int phar_resolve_alias(char *alias, int alias_len, char **filename, int *filename_len);
+PHP_PHAR_API int phar_resolve_alias(char *alias, size_t alias_len, char **filename, size_t *filename_len);
 
 #endif /* PHP_PHAR_H */
